FTR-99 - Refactored QRScanner to use a helper function in the business layer, QRCameraHandler.kt
