+++ app/src/main/java/hr/foi/air/servicesync/ui/components/QRScanner.kt
@@ -1,16 +1,9 @@
 package hr.foi.air.servicesync.ui.components
 
-import android.content.Context
 import android.content.pm.PackageManager
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.annotation.OptIn
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.ExperimentalGetImage
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.camera.view.PreviewView
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -22,26 +15,22 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat
-import com.google.mlkit.vision.barcode.BarcodeScanning
-import com.google.mlkit.vision.common.InputImage
+import hr.foi.air.servicesync.business.setupCamera
 
-
-@OptIn(ExperimentalGetImage::class)
 @Composable
 fun QRScannerContent(modifier: Modifier = Modifier, onCodeScanned: (String) -> Unit) {
     val context = LocalContext.current
     val lifecycleOwner = LocalLifecycleOwner.current
-
-    val cameraProviderFuture = remember { ProcessCameraProvider.getInstance(context) }
     val previewView = remember { PreviewView(context) }
     val permissionLauncher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.RequestPermission(),
         onResult = { isGranted ->
             if (!isGranted) {
-                // Handle permission denied (e.g., show a message or close the screen)
+                Toast.makeText(context, "Camera permission denied", Toast.LENGTH_SHORT).show()
             }
         }
     )
+
     Box(modifier = modifier.fillMaxSize()) {
         AndroidView(
             factory = { previewView },
@@ -49,63 +38,16 @@ fun QRScannerContent(modifier: Modifier = Modifier, onCodeScanned: (String) -> U
         )
 
         LaunchedEffect(Unit) {
-
-            if (ContextCompat.checkSelfPermission(context, android.Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
+            if (ContextCompat.checkSelfPermission(
+                    context,
+                    android.Manifest.permission.CAMERA
+                ) != PackageManager.PERMISSION_GRANTED
+            ) {
                 permissionLauncher.launch(android.Manifest.permission.CAMERA)
-                return@LaunchedEffect // Exit if permission is not granted
+                return@LaunchedEffect
             }
 
-            val cameraProvider = cameraProviderFuture.get()
-            val preview = androidx.camera.core.Preview.Builder().build().also {
-                it.setSurfaceProvider(previewView.surfaceProvider)
-            }
-
-            val barcodeScanner = BarcodeScanning.getClient()
-            val imageAnalyzer = androidx.camera.core.ImageAnalysis.Builder().build().also { imageAnalysis ->
-                imageAnalysis.setAnalyzer(ContextCompat.getMainExecutor(context)) { imageProxy ->
-                    val mediaImage = imageProxy.image
-                    if (mediaImage != null) {
-                        val image = InputImage.fromMediaImage(mediaImage, imageProxy.imageInfo.rotationDegrees)
-                        barcodeScanner.process(image)
-                            .addOnSuccessListener { barcodes ->
-                                for (barcode in barcodes) {
-                                    barcode.rawValue?.let { rawValue ->
-                                        if (rawValue.startsWith("ServiceSync::")) {
-                                            val remainingCode = rawValue.removePrefix("ServiceSync::")
-                                            imageProxy.close()
-                                            onCodeScanned(remainingCode)
-                                            return@addOnSuccessListener
-                                        } else {
-                                            Toast.makeText(
-                                                context,
-                                                "Invalid QR Code",
-                                                Toast.LENGTH_SHORT
-                                            ).show()
-                                        }
-                                    }
-                                }
-                            }
-                            .addOnFailureListener {
-                                // Handle errors if needed
-                            }
-                            .addOnCompleteListener {
-                                imageProxy.close()
-                            }
-                    }
-                }
-            }
-
-            try {
-                cameraProvider.unbindAll()
-                cameraProvider.bindToLifecycle(
-                    lifecycleOwner,
-                    androidx.camera.core.CameraSelector.DEFAULT_BACK_CAMERA,
-                    preview,
-                    imageAnalyzer
-                )
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
+            setupCamera(context, lifecycleOwner, previewView, onCodeScanned)
         }
     }
 }
